health bar

+++ 2dpinball programming/lecture5_samples+/assignment 5.3/Level.cs
@@ -32,7 +32,7 @@ namespace GXPEngine
 		Flipper _player2;
 
 		Players _touched = Players.none;
-		int Damage = 0;
+		float Damage = 0;
